+++ SpeedyTransforms/src/SpeedyTransforms.jl
@@ -1,6 +1,7 @@
 module SpeedyTransforms
 
 using DocStringExtensions, Printf
+import ProgressMeter
 
 # NUMERICS
 import AssociatedLegendrePolynomials as Legendre
@@ -27,6 +28,8 @@ export SpectralTransform,
     transform!,
     transform
 
+export MatrixSpectralTransform
+
 # ALIASING
 export get_nlat_half
 
@@ -51,7 +54,9 @@ export power_spectrum
 include("aliasing.jl")
 include("legendre_shortcuts.jl")
 include("scratch_memory.jl")
+include("gradient_arrays.jl")
 include("spectral_transform.jl")
+include("matrix_transform.jl")
 include("fourier.jl")
 include("legendre.jl")
 include("legendre_ka.jl")